requirements.txt
openai>=1.0
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
tqdm>=4.65
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
lightgbm>=4.0
duckdb>=0.9
geopandas>=0.14
pyarrow>=14.0
umap-learn>=0.5
shap>=0.43
python-dotenv>=1.0
